Remove extra folders
