--- src/models/mod.rs
@@ -1,7 +1,0 @@
-pub mod case;
-pub mod case_notification;
-pub mod civil_case;
-pub mod court_order;
-pub mod filing_deadline;
-pub mod hearing_schedule;
-pub mod recent_case_activity;
