--- src/presentation/api/mod.rs
@@ -1,2 +1,0 @@
-// pub mod user;
-// pub use user::*;
